Refactoring and bugfixing
