+++ src/main/java/Interactors/GameLogic.java
@@ -1,32 +1,30 @@
 package Interactors;
 
 import Entities.*;
+import com.sun.tools.javac.Main;
 import org.hamcrest.core.IsInstanceOf;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
 
 public class GameLogic {
 
-    Player currentPlayer;
-    Board board;
     GameLogicTree[] trees = new GameLogicTree[3];
     GameLogicTree currentTree;
-    ArrayList<Integer> selectedOptions = new ArrayList<Integer>();
-
-    private int[] tradingStates = new int[5];
-    private int[] auctionStates = new int[5];
-    Player[] players;
-
-    //add a method to move the player to a specific cell
+    AuctionTreeHandler auctionTreeHandler;
+    MainTreeHandler mainTreeHandler;
+    TradingTreeHandler tradingTreeHandler;
 
     public GameLogic(Player currentPlayer, Board board){
-        players = board.getPlayers();
-        this.currentPlayer = currentPlayer;
-
-        this.board = board;
+        mainTreeHandler = new MainTreeHandler();
+        auctionTreeHandler = new AuctionTreeHandler(board.getPlayers().length);
+        tradingTreeHandler = new TradingTreeHandler();
+        mainTreeHandler.setGameLogicInteractor(this);
+        mainTreeHandler.initialize(currentPlayer,board);
         createTrees();
     }
+
     public void createTrees(){
         //Creating the game loop tree
         GameLogicTree main = new GameLogicTree("MainTree");
@@ -111,15 +109,6 @@ public class GameLogic {
 
         currentTree = main;
     }
-
-    public void setCurrentPlayer(Player player){
-        this.currentPlayer = player;
-    }
-
-    public Player getCurrentPlayer(){
-        return currentPlayer;
-    }
-
     public State performInput(int input){
         //Moving through the tree depending on the input and the node
         if (input == -1){
@@ -137,22 +126,16 @@ public class GameLogic {
 
         return handleTree(input);
     }
-    public State getInitialState(){
-        State currentState = new State();
-        for (MenuTree tree: currentTree.getChildren()){
-            currentState.addOptions(tree.getName());
-        }
-        return currentState;
-    }
+
     public State handleTree(int input){
-        if (currentTree == trees[0]){
-            return handleMainTree(input);
+        if (currentTree.getMaxParent() == trees[0]){
+            return mainTreeHandler.handleInput(input);
         }
-        else if (currentTree == trees[1]){
-            return handleTradingTree(input);
+        else if (currentTree.getMaxParent() == trees[1]){
+            return tradingTreeHandler.handleInput();
         }
         else{
-            return handleAuctionTree(input);
+            return auctionTreeHandler.handleInput();
         }
 
     }
@@ -174,249 +157,5 @@ public class GameLogic {
         return currentState;
     }
      */
-
-    public State handleTradingTree(int input){
-        State currentState = new State();
-        switch (currentTree.getName()){
-            case "AcceptTrade":
-                Player firstTrader = this.board.getPlayers()[selectedOptions.get(1)];
-                Player secondTrader = this.currentPlayer;
-                Property secondTraderProperty = secondTrader.properties.get(selectedOptions.get(2));
-                Property firstTraderProperty = firstTrader.properties.get(selectedOptions.get(3));
-                // swap the asset owners
-                secondTraderProperty.setOwner(firstTrader);
-                firstTraderProperty.setOwner(secondTrader);
-                secondTrader.properties.remove(secondTraderProperty);
-                secondTrader.properties.add(firstTraderProperty);
-                firstTrader.properties.remove(firstTraderProperty);
-                firstTrader.properties.add(secondTraderProperty);
-                this.currentPlayer = firstTrader;
-                this.currentTree = this.trees[0];
-                break;
-            case "DeclineTrade":
-                firstTrader = this.board.getPlayers()[selectedOptions.get(1)];
-                this.currentPlayer = firstTrader;
-                this.currentTree = this.trees[0];
-                break;
-        }
-        return currentState;
-    }
-    public State handleAuctionTree(int input){
-        State currentState = new State();
-
-        return currentState;
-    }
-    public State handleMainTree(int input){
-        State currentState = new State();
-        switch (currentTree.getName()){
-            case "MainTree":
-                return getInitialState();
-            case "Trade":
-                //Case trade selected
-                //provide a list of all possible players considering the current player is not an option
-                ArrayList<Player> playerCopy = new ArrayList<Player>(Arrays.asList(this.board.getPlayers()));
-                playerCopy.remove(this.currentPlayer);
-                for(int i = 0; i < playerCopy.size(); i++){
-                    currentState.addOptions(Integer.toString(i));
-                }
-                break;
-            case "PickPlayer":
-                //Case player picked
-                //adds the chosen player index in selected options
-                selectedOptions.add(input);
-                //provide item options from the inventory of the selected player
-                Player selectedPlayer = board.getPlayers()[input];
-                ArrayList<Property> playerProperties = selectedPlayer.properties;
-                //using "i" starting from 0 to number of properties the player has - 1
-                for(int i = 0; i < playerProperties.size(); i++){
-                    currentState.addOptions(Integer.toString(i));
-                }
-                break;
-            case "PickItemOp":
-                //Case picking the item of the opponent
-                //the input corresponds to the index of the target player in this.board.getPlayers()
-                selectedOptions.add(input);
-                //provide item options from the current player's inventory
-                ArrayList<Property> currentPlayerInventory = this.currentPlayer.properties;
-                //using "i" starting from 0 to number of properties the player has - 1
-                for(int i = 0; i < this.currentPlayer.properties.size(); i++){
-                    currentState.addOptions(Integer.toString(i));
-                }
-                break;
-            case "PickItemSelf":
-                //the input corresponds to the index of the opponent targeted property;
-                this.selectedOptions.add(input);
-                //Case picking the item of the player
-                //send the trade offer using selectedOptions. Index 0 will be the selected item from opponent and
-                //index 1 will be the selected item from the current player.
-                break;
-            case "SendTrade":
-                this.selectedOptions.add(input);
-                //input corresponds to the index of the current player's selected property
-                //Case sending the trade
-                //the input should be 0 or 1. 0 if the trade was accepted, 1 if the trade was declined.
-                if(input == 0){
-                    //TODO: Process the trade.
-                    int indexOriginalPlayer;
-                    for(int i = 0; i < this.board.getPlayers().length; i++){
-                        if (this.board.getPlayers()[i] == this.currentPlayer){
-                            this.tradingStates[0] = i;
-                            //tradingStates[0] will hold the original player index in this.board.getPlayers()
-                        }
-                    }
-                    this.currentPlayer = this.board.getPlayers()[this.selectedOptions.get(0)];
-                    this.currentTree = this.trees[1];
-                    currentState.addOptions(Integer.toString(0));
-                    currentState.addOptions(Integer.toString(1));
-                }
-                currentState.setEndNode(true);
-                break;
-            case "ManageProperty":
-                //Case manage property selected
-                //provide options on the properties available
-                ArrayList<Property> currentPlayerProperties = this.currentPlayer.properties;
-                for(int i = 0; i < currentPlayerProperties.size(); i++){
-                    currentState.addOptions(Integer.toString(i));
-                }
-                break;
-            case "SelectProperty":
-                //Case property selected (adds the property index)
-                selectedOptions.add(input);
-                //the player chooses what to do to the property
-                //0 is for the mortgage option
-                //1 is for the un mortgaged option
-                //2 is for the build house option
-                currentState.addOptions(Integer.toString(0));
-                currentState.addOptions(Integer.toString(1));
-                currentState.addOptions(Integer.toString(2));
-                break;
-            case "Mortgage":
-                //TODO: finish case 8
-                //Case player selected what to do with the property
-                //the player chooses to mortgage the property
-                Property targetProperty = this.board.getProperties()[selectedOptions.get(0)];
-                this.currentPlayer.mortgage(targetProperty);
-                currentState.setEndNode(true);
-                break;
-            case "BuildProperty":
-                //the player chooses to build houses
-                targetProperty = this.board.getProperties()[selectedOptions.get(0)];
-                this.currentPlayer.buildHouse(targetProperty);
-                currentState.setEndNode(true);
-            case "Roll":
-                //Case roll selected
-                //We can determine if a player lands on a property by checking if the position of
-                //the player is on one with a property on it (not 0,2,7,10,17,20,22,30,33,36,38).
-                Cell landedOnCell = board.getCell(this.currentPlayer.position);
-                if(landedOnCell instanceof Property &&
-                        ((Property) landedOnCell).getOwner() == null) {
-                    currentState.setOnProperty(true);
-                }
-                else {
-                    landedOnCell.performAction(currentPlayer);
-                }
-                break;
-            case "Buy":
-                //Case buy selected
-                //player buys the property that the player lands on
-                targetProperty = board.getProperties()[this.currentPlayer.position];
-                if(this.currentPlayer.money >= targetProperty.getPrice()){
-                    //indicates that the player can afford it and sets the property owner as the current player and
-                    //deducts the player's money.
-                    this.currentPlayer.pay(targetProperty.getPrice());
-                    this.currentPlayer.properties.add(targetProperty);
-                    targetProperty.setOwner(this.currentPlayer);
-                    //1 indicates the property was bought successfully
-                    currentState.addOptions("1");
-                }
-                else{
-                    //0 indicates it failed cause the player has insufficient funds
-                    currentState.addOptions("0");
-                }
-                currentState.setEndNode(true);
-                break;
-            case "Auction":
-                //Case auction selected
-                //TODO: process the auction
-                break;
-            case "Steal":
-                //Case steal selected
-                //provide options of which players we can steal from
-                playerCopy = new ArrayList<Player>(Arrays.asList(this.board.getPlayers()));
-                playerCopy.remove(this.currentPlayer);
-                for(int i = 0; i < playerCopy.size(); i++){
-                    currentState.addOptions(Integer.toString(i));
-                }
-                break;
-            case "ChoosePlayer":
-                //Case player to steal from selected
-                //steal from the player
-                playerCopy = new ArrayList<Player>(Arrays.asList(this.board.getPlayers()));
-                playerCopy.remove(this.currentPlayer);
-                String stealStatus = this.currentPlayer.steal(playerCopy.get(input));
-                currentState.setEndNode(true);
-                break;
-            case "EndTurn":
-                /*TODO: Change the turn, do not let player end turn if
-                negative money
-                * */
-
-                break;
-            case "SettingsMenu":
-                //TODO: show the settings menu
-                break;
-            case "ExitGame":
-                currentState.setExitToMenu(true);
-                break;
-            case "SaveGame":
-                currentState.setSaveGame(true);
-                break;
-            case "Bankruptcy":
-                //TODO: remove player from game and all of their assets
-                currentPlayerProperties = this.currentPlayer.properties;
-                for (Property targetedProperty : currentPlayerProperties) {
-                    targetedProperty.setOwner(null);
-                    targetedProperty.setHouses(0);
-                    targetedProperty.setMortgageStatus(false);
-                }
-                this.board.removePlayer(this.currentPlayer);
-                break;
-        }
-
-        return currentState;
-    }
-
-    public Object[][] playersToArray(){
-        Object[][] playersArray = new Object[players.length][6];
-        for(int i = 0; i < players.length; i++){
-            playersArray[i][0] = players[i].name;
-            playersArray[i][1] = players[i].money;
-            playersArray[i][2] = players[i].properties;
-            playersArray[i][3] = players[i].inJail;
-            playersArray[i][4] = players[i].jailCards;
-            playersArray[i][5] = players[i].position;
-        }
-        return playersArray;
-    }
-
-    public Object[] boardToArray(){
-        Object[] boardArray = new Object[4];
-        boardArray[0] = board.getPlayers();
-        boardArray[1] = board.getCells();
-        boardArray[2] = board.getPlayerPositions();
-        boardArray[3] = board.getProperties();
-        return boardArray;
-    }
-
-    public void movePlayer(int cell_number){
-        int total_squares = 40;
-        int current_player_position = this.currentPlayer.position;
-        if (cell_number - current_player_position < 0) {
-            this.currentPlayer.money = this.currentPlayer.money + 200;
-        }
-        this.currentPlayer.position = cell_number;
-        this.board.updatePlayerPosition(this.currentPlayer);
-    }
-
 }
